+++ src/js/RevealAnimation.js
@@ -111,6 +111,4 @@ const RevealAnimation = {
   },
 };
 
-RevealAnimation.init();
-
 export default RevealAnimation;
